+++ web/components/commands/documentation/AddDocumentation.tsx
@@ -3,9 +3,8 @@ import { Combobox, Dialog, Transition } from '@headlessui/react'
 import { classNames } from '../../../helpers/functions'
 import { DocumentationTypeIcon } from '../../../helpers/Icons'
 import { ChevronRightIcon } from '@heroicons/react/solid'
-import { Org } from '../../../pages'
 import DocumentationConfig from './DocumentationConfig'
-import { User } from '../../../context/UserContex'
+import { Org, User } from '../../../context/ProfileContex'
 
 export type AddDocumentationType = 'webpage' | 'notion' | 'googledocs' | 'confluence';
